chore: huge refactoring
