--- command_version.go
@@ -1,6 +1,0 @@
-package main
-
-func Version() Response {
-	return Success(version)
-}
-
