+++ commands/deploy.go
@@ -1,9 +1,12 @@
-package main
+package commands
 
 import (
 	"bytes"
 	"errors"
 	"fmt"
+	"github.com/sachamorard/swapper/response"
+	"github.com/sachamorard/swapper/utils"
+	"github.com/sachamorard/swapper/yaml"
 	"io/ioutil"
 	"net/http"
 	"strings"
@@ -38,44 +41,44 @@ Examples:
 )
 
 func DeployArgs(argv []string) docopt.Opts {
-	hostname, _ := GetHostname()
+	hostname, _ := utils.GetHostname()
 	deployUsage = strings.Replace(deployUsage, "{{hostname}}", hostname, -1)
 
 	arguments, _ := docopt.ParseArgs(deployUsage, argv, "")
 	return arguments
 }
 
-func Deploy(argv []string) Response {
+func Deploy(argv []string) response.Response {
 	arguments := DeployArgs(argv)
-	vars := InterfaceToArray(arguments["--var"])
+	vars := utils.InterfaceToArray(arguments["--var"])
 	file := arguments["--file"].(string)
 	masterHostname := arguments["--master"].(string)
-	cleanYaml, err := PrepareSwapperYaml(file, vars)
+	cleanYaml, err := yaml.PrepareSwapperYaml(file, vars)
 	if err != nil {
-		return Fail(err.Error())
+		return response.Fail(err.Error())
 	}
 
-	yamlConf, err := ParseSwapperYaml(cleanYaml)
+	yamlConf, err := yaml.ParseSwapperYaml(cleanYaml)
 	if err != nil {
-		return Fail(err.Error())
+		return response.Fail(err.Error())
 	}
 	if len(yamlConf.Masters) != 0 {
-		return Fail(errorMessages["no_masters_field"])
+		return response.Fail(response.ErrorMessages["no_masters_field"])
 	}
 
 	if yamlConf.Time != 0 {
-		return Fail(errorMessages["no_time_field"])
+		return response.Fail(response.ErrorMessages["no_time_field"])
 	}
 
 	if yamlConf.Hash != "" {
-		return Fail(errorMessages["no_hash_field"])
+		return response.Fail(response.ErrorMessages["no_hash_field"])
 	}
 
 	if masterHostname == "localhost" {
-		masterHostname, _ = GetHostname()
+		masterHostname, _ = utils.GetHostname()
 	}
 	if masterHostname == "127.0.0.1" {
-		masterHostname, _ = GetHostname()
+		masterHostname, _ = utils.GetHostname()
 	}
 
 	i := strings.Index(masterHostname, ":")
@@ -85,7 +88,7 @@ func Deploy(argv []string) Response {
 
 	swapperYml := CurlSwapperYaml(masterHostname)
 	if swapperYml == "" {
-		return Fail(fmt.Sprintf(errorMessages["bad_master_addr"], masterHostname))
+		return response.Fail(fmt.Sprintf(response.ErrorMessages["bad_master_addr"], masterHostname))
 	}
 
 	masterSplit := strings.Split(masterHostname, ":")
@@ -94,32 +97,32 @@ func Deploy(argv []string) Response {
 	return DeployFile(cleanYaml, port, yamlConf)
 }
 
-func DeployFile(cleanYaml string, port string, yamlConf YamlConf) Response {
+func DeployFile(cleanYaml string, port string, yamlConf yaml.YamlConf) response.Response {
 	err := DeployReq(cleanYaml, port)
 	if err != nil {
-		_ = SlackSendError("Deployment failed\n"+err.Error(), yamlConf)
-		return Fail(err.Error())
+		_ = utils.SlackSendError("Deployment failed\n"+err.Error(), yamlConf)
+		return response.Fail(err.Error())
 	}
-	_ = SlackSendSuccess("Deployment succeed", yamlConf)
-	return Success("\n>> Deployment succeed\n")
+	_ = utils.SlackSendSuccess("Deployment succeed", yamlConf)
+	return response.Success("\n>> Deployment succeed\n")
 }
 
 func DeployReq(cleanYaml string, port string) (err error) {
 	req, err := http.NewRequest(http.MethodPost, "http://localhost:"+port+"/swapper.yml", bytes.NewBuffer([]byte(cleanYaml)))
 	if err != nil {
-		return errors.New(fmt.Sprintf(errorMessages["request_failed"], err))
+		return errors.New(fmt.Sprintf(response.ErrorMessages["request_failed"], err))
 	}
 	req.Header.Set("Content-Type", "text/yml")
 	client := &http.Client{Timeout: 5 * time.Second}
 	resp, err := client.Do(req)
 	if err != nil {
-		return errors.New(fmt.Sprintf(errorMessages["deploy_failed"], err))
+		return errors.New(fmt.Sprintf(response.ErrorMessages["deploy_failed"], err))
 	}
 	defer resp.Body.Close()
 
 	if resp.Status != "200 OK" {
 		body, _ := ioutil.ReadAll(resp.Body)
-		return errors.New(fmt.Sprintf(errorMessages["deploy_failed"], body))
+		return errors.New(fmt.Sprintf(response.ErrorMessages["deploy_failed"], body))
 	} else {
 		return nil
 	}
