+++ yaml/tests/v1/swapper.invalid.0.yml
