+++ yaml/tests/v1/swapper.invalid.6.yml
