+++ yaml/tests/v1/swapper.invalid.7.yml
