+++ yaml/tests/v1/swapper.invalid.yml
